Develop

+++ src/main/java/org/example/controller/ReviewController.java
@@ -2,6 +2,9 @@ package org.example.controller;
 
 import org.example.entity.Post;
 import org.example.entity.Review;
+import org.example.entity.User;
+import org.example.model.PostDTO;
+import org.example.model.ReviewDTO;
 import org.example.repository.PostRepository;
 import org.example.repository.ReviewRepository;
 import org.example.repository.UserRepository;
@@ -34,8 +37,11 @@ public class ReviewController {
     public ResponseEntity<List<Review>> getAllByUser(){
 
         try {
-            // first, take current user's ID, then get user from database and put into getAllReservationsByUser
-            ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.getAllReviewsByUser(this.userRepository.getById(this.identity.getCurrent().getId())));
+            // Authorize user
+            User user = identity.getCurrent();
+            if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+            ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.getAllReviewsByUser(user));
 
             if (reviews.isEmpty()) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             return new ResponseEntity<>(reviews, HttpStatus.OK);
@@ -44,12 +50,39 @@ public class ReviewController {
         }
     }
 
-//    @GetMapping("/")
-//    public ResponseEntity<List<Review>> getAllByPost(){
+//    @GetMapping("/post/{title}")
+//    public ResponseEntity<List<Review>> getAllByPostTitle(@PathVariable("title") String title){
 //
 //        try {
-//            // first, take current user's ID, then get user from database and put into getAllReservationsByUser
-//            ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.getAllReviewsByPost(this.postRepository.findAllByUserId(this.identity.getCurrent().getId())));
+//            // Authorize user
+//            User user = identity.getCurrent();
+//            if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+//
+//            List<Post> reviews = new ArrayList<>(this.postRepository.getAllByUser(user));
+//
+//            for (Post element : tempList) {
+//                if(element.getTitle().equals(title)){
+//                    reviews = new ArrayList<>(this.reviewRepository.getAllReviewsByPost(element));
+//                    if (reviews.isEmpty()) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+//                    return new ResponseEntity<>(reviews, HttpStatus.OK);
+//                }
+//            }
+//
+//            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+//        } catch(Exception e) {
+//            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+//        }
+//    }
+
+//    @GetMapping("/review/{title}")
+//    public ResponseEntity<List<Review>> getAllReviewsByTitle(@PathVariable("title") String title){
+//
+//        try {
+//            // Authorize user
+//            User user = identity.getCurrent();
+//            if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+//
+//            ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.getAllReviewsByTitle(title));
 //
 //            if (reviews.isEmpty()) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 //            return new ResponseEntity<>(reviews, HttpStatus.OK);
@@ -58,29 +91,19 @@ public class ReviewController {
 //        }
 //    }
 
-    @GetMapping("/title/{title}")
-    public ResponseEntity<List<Review>> getAllReviewsByTitle(@PathVariable("title") String title){
 
-        try {
-            // first, take current user's ID, then get user from database and put into getAllReservationsByUser
-            ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.getAllReviewsByTitle(title));
-
-            if (reviews.isEmpty()) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-            return new ResponseEntity<>(reviews, HttpStatus.OK);
-        } catch(Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-
-    @GetMapping("/post/{postId}/likes")
+    @GetMapping("/post/likes/{postId}")
     public ResponseEntity<List<Review>> findAllReviewsByPostOrderByLikesDesc(@PathVariable("postId") UUID postId){
 
         try {
-            // first, take current user's ID, then get user from database and put into getAllReservationsByUser
+            // Authorize user
+            User user = identity.getCurrent();
+            if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
             ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.findAllReviewsByPostOrderByLikesDesc(postId));
 
             if (reviews.isEmpty()) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+
             return new ResponseEntity<>(reviews, HttpStatus.OK);
         } catch(Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -88,14 +111,16 @@ public class ReviewController {
     }
 
 
-    @GetMapping("/post/{postId}/dislikes")
+    @GetMapping("/post/dislikes/{postId}")
     public ResponseEntity<List<Review>> findAllReviewsByPostOrderByDislikesDesc(@PathVariable("postId") UUID postId){
 
         try {
-            // first, take current user's ID, then get user from database and put into getAllReservationsByUser
-            ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.findAllReviewsByPostOrderByDislikesDesc(postId));
+            // Authorize user
+            User user = identity.getCurrent();
+            if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
 
-            if (reviews.isEmpty()) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            ArrayList<Review> reviews = new ArrayList<>(this.reviewRepository.findAllReviewsByPostOrderByDislikesDesc(postId));
+            if (reviews.isEmpty()) return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
             return new ResponseEntity<>(reviews, HttpStatus.OK);
         } catch(Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -103,11 +128,15 @@ public class ReviewController {
     }
 
 
-    @GetMapping("/{reservationId}")
-    public ResponseEntity<Review> getById(@PathVariable("reservationId") UUID reviewId){
+    @GetMapping("/{reviewId}")
+    public ResponseEntity<Review> getById(@PathVariable("reviewId") UUID reviewId){
         try{
-            Review review = this.reviewRepository.getReviewByUserAndId(this.userRepository.getById(this.identity.getCurrent().getId()), reviewId);
-            if(review == null) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            // Authorize user
+            User user = identity.getCurrent();
+            if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+            Review review = this.reviewRepository.getReviewByUserAndId(user, reviewId);
+            if(review == null) return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
             return new ResponseEntity<>(review, HttpStatus.OK);
         } catch(Exception e){
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -115,25 +144,52 @@ public class ReviewController {
     }
 
     @PostMapping("/")
-    public ResponseEntity<?> save(@RequestBody Review review){
+    public ResponseEntity<?> save(@RequestBody ReviewDTO reviewDTO){
+
+        // Authorize user
+        User user = identity.getCurrent();
+        if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        Post post = this.postRepository.getById(reviewDTO.getPost().getId());
+
         try{
+            Review review = this.reviewRepository.save(Review.builder().
+                    post(post).user(user).
+                    title(reviewDTO.getTitle()).
+                    contents(reviewDTO.getContents()).
+                    exposureDate(reviewDTO.getExposureDate()).
+                    likes(reviewDTO.getLikes()).
+                    dislikes(reviewDTO.getDislikes()).build());
+
             this.reviewRepository.save(review);
-            return new ResponseEntity<>(HttpStatus.CREATED);
+            return new ResponseEntity<>(review, HttpStatus.CREATED);
         } catch(Exception e){
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-
     @PutMapping("/{reviewId}")
-    public ResponseEntity<?> updateById(@PathVariable("reviewId") UUID reviewId, @RequestBody Review review){
+    public ResponseEntity<?> update(@PathVariable("reviewId") UUID reviewId, @RequestBody ReviewDTO reviewDTO){
+        // Authorize user
+        User user = identity.getCurrent();
+        if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        Post post = this.postRepository.getById(reviewDTO.getPost().getId());
+
         try{
-            Review currentReview = this.reviewRepository.getReviewByUserAndId(this.userRepository.getById(this.identity.getCurrent().getId()), reviewId);
-            if(currentReview == null) return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
-            review.setId(reviewId);
-            this.reviewRepository.deleteById(reviewId);
-            this.reviewRepository.save(review);
-            return new ResponseEntity<>(review, HttpStatus.OK);
+            // Delete existing review
+            deleteById(reviewId);
+
+            // Add new ones
+            Review review = this.reviewRepository.save(Review.builder().
+                    post(post).user(user).
+                    title(reviewDTO.getTitle()).
+                    contents(reviewDTO.getContents()).
+                    exposureDate(reviewDTO.getExposureDate()).
+                    likes(reviewDTO.getLikes()).
+                    dislikes(reviewDTO.getDislikes()).build());
+
+            return new ResponseEntity<>(review,HttpStatus.CREATED);
         } catch (Exception e){
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -142,6 +198,10 @@ public class ReviewController {
 
     @DeleteMapping("/")
     public ResponseEntity<?> deleteAll(){
+        // Authorize user
+        User user = identity.getCurrent();
+        if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         try{
             this.reviewRepository.deleteAllReviewsByUser(this.userRepository.getById(this.identity.getCurrent().getId()));
             return new ResponseEntity<>(null, HttpStatus.OK);
@@ -153,9 +213,14 @@ public class ReviewController {
 
     @DeleteMapping("/{reviewId}")
     public ResponseEntity<?> deleteById(@PathVariable("reviewId") UUID reviewId){
+        // Authorize user
+        User user = identity.getCurrent();
+        if (user == null) return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         try{
-            Review currentReview = this.reviewRepository.getReviewByUserAndId(this.userRepository.getById(this.identity.getCurrent().getId()), reviewId);
-            if(currentReview == null) return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+            Review review = this.reviewRepository.getReviewByUserAndId(user, reviewId);
+            if(review == null) return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+
             this.reviewRepository.deleteById(reviewId);
             return new ResponseEntity<>(null, HttpStatus.OK);
         } catch (Exception e){
